untested but at least docs compiling version
