+++ src/xmris/__init__.py
@@ -1,32 +1,36 @@
 # 0. Expose the submodules for quartodoc (Griffe) traversal
-from . import accessor, config, fitting, processing, utils, vendor  # noqa: I001
+from . import fitting, processing, vendor  # noqa: I001
 
 # 1. Configuration (The Central Nervous System)
-from .config import DEFAULTS
+from .core import ATTRS, COORDS, DIMS, VARS, accessor
 
 # 2. Accessor (Importing this automatically registers the .xmr namespace)
-from .accessor import XmrisAccessor, XmrisDatasetAccessor
+from .core.accessor import XmrisAccessor, XmrisDatasetAccessor
 
-# 3. Utilities
-from .utils import to_complex, to_real_imag
+# legacy -- patch for now (branch refactor/18)
+from .config import DEFAULTS
+
+# 6. Fitting
+from .fitting.amares import fit_amares
+from .processing import utils
 
 # 4. Core Processing
 from .processing.fid import apodize_exp, apodize_lg, to_fid, to_spectrum, zero_fill
 from .processing.fourier import fft, fftc, fftshift, ifft, ifftc, ifftshift
 from .processing.phase import autophase, phase
 
+# 3. Utilities
+from .processing.utils import to_complex, to_real_imag
+
 # 5. Vendor Specific
 from .vendor.bruker import remove_digital_filter
-
-# 6. Fitting
-from .fitting.amares import fit_amares
 
 # 7. Visualisation
 from .visualization.plot import (
     PlotHeatmapConfig,
+    PlotQCGridConfig,
     PlotRidgeConfig,
     PlotTrajectoryConfig,
-    PlotQCGridConfig,
 )
 
 # Explicitly define the public API.
@@ -40,7 +44,11 @@ __all__ = [
     "vendor",
     # --- Flat API (For the users) ---
     # Config
-    "DEFAULTS",
+    "ATTRS",
+    "COORDS",
+    "DIMS",
+    "VARS",
+    "DEFAULTS",  # legacy -- patch for now (branch refactor/18)
     # Accessor
     "XmrisAccessor",
     "XmrisDatasetAccessor",
